fix(experience): smart input field should have correct initial type
